investment_tracker/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
Django==3.2.5
factory-boy==3.2.0
Faker==8.10.3
mypy-extensions==0.4.3
pathspec==0.9.0
psycopg2==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
tomli==1.2.0
